+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Updated/Enhanced Rules and Bug Fixes
    - #314: **VariableTypeRequired** rule: Add *ignoreVariableNames* property.
    - #314: **FieldTypeRequiredRule** rule: Add *ignoreFieldNames* property.
    - #235: **UnnecessaryGetter**: Ignore getters within calls to Spock `Stub()`/`Mock()`. 
+   - #157 **SpaceAroundOperatorAstVisitor**: Check for space around equals for declaration expressions in variables and fields.
 
 Framework and Infrastructure
    - #311: Add equivalent linux command line example. (Wilfred Hughes)